routing: divide main()
