+++ routing/Makefile
@@ -1,16 +1,21 @@
-SRC = routing.cpp
+SRC = ln_routing.cpp
 OBJ = routing
 CFLAGS = -I../include -I../ucoin/include -I../libs/install/include -I../ucoin/libs/install/include
 LDFLAGS = -L../libs/install/lib -L../ucoin/libs/install/lib -L../ucoin
 LDFLAGS += -linih -llmdb -lucoin -lsodium -lbase58 -lmbedcrypto -pthread
 
-all: routing
+all: lnrouting routing
 
 routing: $(SRC)
-	g++ -W -Wall -ggdb $(CFLAGS) -o $(OBJ) $(SRC) ../cmn/conf.c ../cmn/misc.c $(LDFLAGS)
+	gcc -c -W -Wall -ggdb $(CFLAGS) ../cmn/misc.c
+	gcc -c -W -Wall -ggdb $(CFLAGS) routing_main.c
+	g++ -o $(OBJ)  -W -Wall -ggdb $(CFLAGS) ln_routing.o misc.o routing_main.o $(LDFLAGS) -lgcc
+
+lnrouting:
+	g++ -c -W -Wall -ggdb $(CFLAGS) $(SRC)
 
 clean:
-	-rm -rf $(OBJ) .Depend
+	-rm -rf $(OBJ) *.o .Depend
 
 .Depend:
 ifneq ($(MAKECMDGOALS),clean)
